+++ src/Console/Framework/Configuration/Routing.php
@@ -12,8 +12,8 @@ class Routing{
 	public static function file(): string{
 		$content = '<?php'."\n";
 		$content .= 'namespace Configuration;'."\n";
-		$content .= 'use Frash\Framework\Routing\TreatmentPhp;'."\n\n";
-		$content .= 'class Routing extends TreatmentPhp{'."\n";
+		$content .= 'use Frash\Framework\Routing\Router;'."\n\n";
+		$content .= 'class Routing extends Router{'."\n";
 		$content .= '	public function __construct(){'."\n";
 		$content .= '		'."\n";
 		$content .= '	}'."\n";
